downloading osm data and converting element bounds to meshes

+++ gisgameworld/Assets/_gisgameworld/Scripts/Data/OSM/OSMCoordinate.cs
@@ -3,6 +3,6 @@ using System.Collections;
 
 public class OSMCoordinate
 {
-    public double lat;
-    public double lon;
+    public float lat;
+    public float lon;
 }
